+++ packages/integration-kit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @trigger.dev/integration-kit
 
+## 2.2.4
+
 ## 2.2.3
 
 ## 2.2.2