+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @trigger.dev/nestjs
 
+## 2.2.4
+
+### Patch Changes
+
+- Updated dependencies [c1710ae7]
+- Updated dependencies [9c4be40a]
+  - @trigger.dev/sdk@2.2.4
+
 ## 2.2.3
 
 ### Patch Changes